pyproject.toml
[project]
name = "clickup-agent"
version = "0.1.0"
description = "ClickUp agent driven via claude-telegram-bot (ctb). Targets Oysterun."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "typer>=0.12",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
    "rapidfuzz>=3.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tool_scripts"]
